Return 2nd handshake
On the server, receive the first handshake and return the second. The
socket-swapping is also in place; the client must be able to receive
from a random port. No processing of the client-side random number is
done; the server-side random number is simply appended to the payload
data (the two numbers are separated by a comma).